+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve error message quality in type checker.
 - Refactor CLI interface implementation.
 - Refactor benchmark tests.
+- Improve look-ahead branching heuristic.
 
 ### Fixed
 
@@ -59,7 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support generic type for datatypes and functions.
 - Support file dump in output directory.
-- Implement lookahead branching heuristic.
+- Implement look-ahead branching heuristic.
 - Support interactive debug mode.
 
 ### Changed